Refactor Buffer; use new functionality in NSEC3 digest code

+++ contracts/SHA1NSEC3Digest.sol
@@ -14,12 +14,15 @@ contract SHA1NSEC3Digest is NSEC3Digest {
     buf.append(data);
     buf.append(salt);
     bytes20 h = SHA1.sha1(buf.buf);
-
-    for(uint i = 0; i < iterations; i++) {
+    if(iterations > 0) {
       buf.truncate();
-      buf.append(h);
+      buf.append(bytes20(0));
       buf.append(salt);
-      h = SHA1.sha1(buf.buf);
+
+      for(uint i = 0; i < iterations; i++) {
+        buf.write(0, h);
+        h = SHA1.sha1(buf.buf);
+      }
     }
 
     buf.truncate();
